Archivo OtrosDispositivos

+++ M13_Practica2_MWC/src/main/java/com/mycompany/m13_practica2_mwc/Dispositivos.java
@@ -28,7 +28,7 @@ public class Dispositvos {
         this.model = model;
     }
 
-    public int getPreuBase() {
+    public double getPreuBase() {
         return preuBase;
     }
